[Consignments] Move some tests to jest
